+++ src/main/kotlin/com/jetbrains/snakecharm/codeInsight/completion/SmkKeywordCompletionContributor.kt
@@ -9,7 +9,6 @@ import com.intellij.patterns.PlatformPatterns
 import com.intellij.patterns.PlatformPatterns.psiElement
 import com.intellij.patterns.PsiElementPattern
 import com.intellij.patterns.StandardPatterns
-import com.intellij.profile.codeInspection.InspectionProfileManager
 import com.intellij.psi.PsiComment
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
@@ -18,14 +17,19 @@ import com.intellij.util.ProcessingContext
 import com.jetbrains.python.PyTokenTypes
 import com.jetbrains.python.codeInsight.completion.PythonLookupElement
 import com.jetbrains.python.psi.*
+import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.MODULE_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.RULE_OR_CHECKPOINT_SECTION_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.SUBWORKFLOW_SECTIONS_KEYWORDS
-import com.jetbrains.snakecharm.inspections.SmkUnrecognizedSectionInspection
+import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.USE_DECLARATION_KEYWORDS
+import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.USE_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.lang.SnakemakeLanguageDialect
+import com.jetbrains.snakecharm.lang.SnakemakeNames
 import com.jetbrains.snakecharm.lang.parser.SmkTokenTypes.RULE_LIKE
+import com.jetbrains.snakecharm.lang.parser.SmkTokenTypes.USE_KEYWORD
 import com.jetbrains.snakecharm.lang.parser.SmkTokenTypes.WORKFLOW_TOPLEVEL_DECORATORS_WO_RULE_LIKE
 import com.jetbrains.snakecharm.lang.parser.SnakemakeLexer
 import com.jetbrains.snakecharm.lang.psi.*
+import com.jetbrains.snakecharm.lang.psi.impl.SmkUseArgsSectionImpl
 
 /**
  * @author Roman.Chernyatchik
@@ -54,6 +58,25 @@ class SmkKeywordCompletionContributor : CompletionContributor() {
             SubworkflowSectionKeywordsProvider.CAPTURE,
             SubworkflowSectionKeywordsProvider
         )
+
+        extend(
+            CompletionType.BASIC,
+            ModuleSectionKeywordsProvider.CAPTURE,
+            ModuleSectionKeywordsProvider
+        )
+
+        extend(
+            CompletionType.BASIC,
+            UseSectionKeywordsProvider.CAPTURE,
+            UseSectionKeywordsProvider
+        )
+    }
+}
+
+object RuleKeywordTail : TailType() {
+    override fun processTail(editor: Editor, tailOffset: Int): Int {
+        editor.document.insertString(tailOffset, " ${SnakemakeNames.RULE_KEYWORD} ")
+        return moveCaret(editor, tailOffset, 2 + SnakemakeNames.RULE_KEYWORD.length)
     }
 }
 
@@ -95,7 +118,6 @@ object WorkflowTopLevelKeywordsProvider : CompletionProvider<CompletionParameter
         val tokenType2Name = SnakemakeLexer.KEYWORDS
             .map { (k, v) -> v to k }
             .toMap()
-
         listOf(
             WORKFLOW_TOPLEVEL_DECORATORS_WO_RULE_LIKE to ColonAndWhiteSpaceTail,
             RULE_LIKE to TailType.SPACE
@@ -103,10 +125,12 @@ object WorkflowTopLevelKeywordsProvider : CompletionProvider<CompletionParameter
             tokenSet.types.forEach { tt ->
                 val s = tokenType2Name[tt]!!
 
+                val modifiedTail = if (tt == USE_KEYWORD) RuleKeywordTail else tail
+
                 result.addElement(
                     SmkCompletionUtil.createPrioritizedLookupElement(
                         TailTypeDecorator.withTail(
-                            PythonLookupElement(s, true, null), tail
+                            PythonLookupElement(s, true, null), modifiedTail
                         ),
                         SmkCompletionUtil.KEYWORDS_PRIORITY
                     )
@@ -152,7 +176,10 @@ object RuleSectionKeywordsProvider : CompletionProvider<CompletionParameters>() 
         .inside(SmkRuleOrCheckpoint::class.java)!!
         .andNot(
             psiElement().insideOneOf(
-                PyArgumentList::class.java, SmkRunSection::class.java, PsiComment::class.java
+                PyArgumentList::class.java,
+                SmkRunSection::class.java,
+                PsiComment::class.java,
+                SmkUseArgsSectionImpl::class.java
             )
         )
 
@@ -191,6 +218,73 @@ object SubworkflowSectionKeywordsProvider : CompletionProvider<CompletionParamet
         result: CompletionResultSet
     ) {
         SUBWORKFLOW_SECTIONS_KEYWORDS.forEach { s ->
+            result.addElement(
+                SmkCompletionUtil.createPrioritizedLookupElement(
+                    TailTypeDecorator.withTail(
+                        PythonLookupElement(s, true, PlatformIcons.PROPERTY_ICON),
+                        ColonAndWhiteSpaceTail
+                    ),
+                    priority = SmkCompletionUtil.SECTIONS_KEYS_PRIORITY
+                )
+            )
+        }
+    }
+}
+
+object ModuleSectionKeywordsProvider : CompletionProvider<CompletionParameters>() {
+    val CAPTURE = psiElement()
+        .inFile(SmkKeywordCompletionContributor.IN_SNAKEMAKE)
+        .inside(psiElement().inside(SmkModule::class.java))
+        .andNot(
+            psiElement().insideOneOf(PyArgumentList::class.java, PsiComment::class.java)
+        )
+
+    override fun addCompletions(
+        parameters: CompletionParameters,
+        context: ProcessingContext,
+        result: CompletionResultSet
+    ) {
+        MODULE_SECTIONS_KEYWORDS.forEach { s ->
+            result.addElement(
+                SmkCompletionUtil.createPrioritizedLookupElement(
+                    TailTypeDecorator.withTail(
+                        PythonLookupElement(s, true, PlatformIcons.PROPERTY_ICON),
+                        ColonAndWhiteSpaceTail
+                    ),
+                    priority = SmkCompletionUtil.SECTIONS_KEYS_PRIORITY
+                )
+            )
+        }
+    }
+}
+
+object UseSectionKeywordsProvider : CompletionProvider<CompletionParameters>() {
+    val CAPTURE = psiElement()
+        .inFile(SmkKeywordCompletionContributor.IN_SNAKEMAKE)
+        .inside(psiElement().inside(SmkUse::class.java))
+        .andNot(
+            psiElement().insideOneOf(PyArgumentList::class.java, PsiComment::class.java)
+        )
+
+    override fun addCompletions(
+        parameters: CompletionParameters,
+        context: ProcessingContext,
+        result: CompletionResultSet
+    ) {
+        USE_DECLARATION_KEYWORDS.forEach { s ->
+            val tail = if (s != SnakemakeNames.SMK_WITH_KEYWORD) TailType.SPACE else TailType.CASE_COLON
+            result.addElement(
+                SmkCompletionUtil.createPrioritizedLookupElement(
+                    TailTypeDecorator.withTail(
+                        PythonLookupElement(s, true, PlatformIcons.PROPERTY_ICON),
+                        tail
+                    ),
+                    priority = SmkCompletionUtil.SECTIONS_KEYS_PRIORITY
+                )
+            )
+        }
+
+        USE_SECTIONS_KEYWORDS.forEach { s ->
             result.addElement(
                 SmkCompletionUtil.createPrioritizedLookupElement(
                     TailTypeDecorator.withTail(