+++ src/main/kotlin/com/jetbrains/snakecharm/lang/highlighter/SmkSyntaxAnnotator.kt
@@ -1,11 +1,14 @@
 package com.jetbrains.snakecharm.lang.highlighter
 
+import com.jetbrains.python.PyTokenTypes
 import com.jetbrains.python.highlighting.PyHighlighter
 import com.jetbrains.python.highlighting.PyHighlighter.PY_FUNC_DEFINITION
+import com.jetbrains.snakecharm.lang.parser.SmkTokenTypes
 import com.jetbrains.snakecharm.lang.psi.*
+import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkElementTypes
 import com.jetbrains.snakecharm.lang.validation.SmkAnnotator
 
-object SmkSyntaxAnnotator: SmkAnnotator() {
+object SmkSyntaxAnnotator : SmkAnnotator() {
     override fun visitSmkRule(rule: SmkRule) {
         highlightSMKRuleLike(rule)
     }
@@ -18,11 +21,41 @@ object SmkSyntaxAnnotator: SmkAnnotator() {
         highlightSMKRuleLike(subworkflow)
     }
 
+    override fun visitSmkModule(module: SmkModule) {
+        highlightSMKRuleLike(module)
+    }
+
+    override fun visitSmkUse(use: SmkUse) {
+        highlightWorkflowSection(use)
+
+        val node = use.node.firstChildNode
+        var next = node.treeNext
+        var done = false
+        while (!done && next != null) {
+            when (next.elementType) {
+                SmkTokenTypes.RULE_KEYWORD, SmkTokenTypes.SMK_FROM_KEYWORD,
+                SmkTokenTypes.SMK_AS_KEYWORD, SmkTokenTypes.SMK_WITH_KEYWORD -> addHighlightingAnnotation(
+                    next,
+                    PyHighlighter.PY_KEYWORD
+                )
+                SmkElementTypes.USE_NAME_IDENTIFIER, PyTokenTypes.IDENTIFIER -> addHighlightingAnnotation(
+                    next, PY_FUNC_DEFINITION
+                )
+                PyTokenTypes.COLON -> done = true
+            }
+            next = next.treeNext
+        }
+    }
+
     override fun visitSmkRuleOrCheckpointArgsSection(st: SmkRuleOrCheckpointArgsSection) {
         highlightRuleLikeSection(st)
     }
 
     override fun visitSmkSubworkflowArgsSection(st: SmkSubworkflowArgsSection) {
+        highlightRuleLikeSection(st)
+    }
+
+    override fun visitSmkModuleArgsSection(st: SmkModuleArgsSection) {
         highlightRuleLikeSection(st)
     }