+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/SmkFile.kt
@@ -37,6 +37,36 @@ class SmkFile(viewProvider: FileViewProvider) : PyFileImpl(viewProvider, Snakema
         return subworkflowNameAndPsi
     }
 
+    fun collectModules(): List<Pair<String, SmkModule>> {
+        val moduleNameAndPsi = arrayListOf<Pair<String, SmkModule>>()
+
+        acceptChildren(object : PyElementVisitor(), SmkElementVisitor {
+            override val pyElementVisitor: PyElementVisitor = this
+
+            override fun visitSmkModule(module: SmkModule) {
+                if (module.name != null) {
+                    moduleNameAndPsi.add(module.name!! to module)
+                }
+            }
+        })
+        return moduleNameAndPsi
+    }
+
+    fun collectUses(): List<Pair<String, SmkUse>> {
+        val useNameAndPsi = arrayListOf<Pair<String, SmkUse>>()
+
+        acceptChildren(object : PyElementVisitor(), SmkElementVisitor {
+            override val pyElementVisitor: PyElementVisitor = this
+
+            override fun visitSmkUse(use: SmkUse) {
+                if (use.name != null) {
+                    useNameAndPsi.add(use.name!! to use)
+                }
+            }
+        })
+        return useNameAndPsi
+    }
+
     fun collectCheckPoints(): List<Pair<String, SmkCheckPoint>> {
         val checkpointNameAndPsi = arrayListOf<Pair<String, SmkCheckPoint>>()