+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/SmkPsiElements.kt
@@ -7,9 +7,9 @@ import com.jetbrains.python.codeInsight.controlflow.ScopeOwner
 import com.jetbrains.python.psi.*
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_DEFINING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_EXPANDING_SECTIONS_KEYWORDS
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkCheckpointStub
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkRuleStub
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkSubworkflowStub
+import com.jetbrains.snakecharm.codeInsight.resolve.SmkResolveUtil
+import com.jetbrains.snakecharm.lang.psi.stubs.*
+import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLExpression
 
 interface SmkToplevelSection : SmkSection {
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint? = null
@@ -20,6 +20,10 @@ interface SmkRule : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkRuleStub>
 interface SmkCheckPoint : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkCheckpointStub>
 
 interface SmkSubworkflow : SmkRuleLike<SmkSubworkflowArgsSection>, StubBasedPsiElement<SmkSubworkflowStub>
+
+interface SmkModule : SmkRuleLike<SmkModuleArgsSection>, StubBasedPsiElement<SmkModuleStub>
+
+interface SmkUse : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkUseStub>
 
 interface SmkRuleOrCheckpointArgsSection : SmkArgsSection, PyTypedElement { // PyNamedElementContainer
     /**
@@ -41,6 +45,10 @@ interface SmkRuleOrCheckpointArgsSection : SmkArgsSection, PyTypedElement { // P
 }
 
 interface SmkSubworkflowArgsSection : SmkArgsSection {
+    override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint? = null
+}
+
+interface SmkModuleArgsSection : SmkArgsSection {
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint? = null
 }