+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/SmkUseImpl.kt
@@ -1,0 +1,32 @@
+package com.jetbrains.snakecharm.lang.psi.impl
+
+import com.intellij.lang.ASTNode
+import com.jetbrains.python.psi.PyElementType
+import com.jetbrains.python.psi.PyElementVisitor
+import com.jetbrains.python.psi.types.TypeEvalContext
+import com.jetbrains.snakecharm.lang.parser.SmkTokenTypes
+import com.jetbrains.snakecharm.lang.psi.SmkElementVisitor
+import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpointArgsSection
+import com.jetbrains.snakecharm.lang.psi.SmkUse
+import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkStubElementTypes
+import com.jetbrains.snakecharm.lang.psi.stubs.SmkUseStub
+import com.jetbrains.snakecharm.lang.psi.types.SmkRuleLikeSectionType
+
+class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSection>, SmkUse {
+
+    constructor(node: ASTNode) : super(node)
+    constructor(stub: SmkUseStub) : super(stub, SmkStubElementTypes.USE_DECLARATION_STATEMENT)
+
+    override val sectionTokenType: PyElementType = SmkTokenTypes.USE_KEYWORD
+
+    override val wildcardsElement = SmkRuleImpl.createFakeWildcardsPsiElement(this)
+
+    override fun getType(context: TypeEvalContext, key: TypeEvalContext.Key) = SmkRuleLikeSectionType(this)
+
+    override fun acceptPyVisitor(pyVisitor: PyElementVisitor?) {
+        when (pyVisitor) {
+            is SmkElementVisitor -> pyVisitor.visitSmkUse(this)
+            else -> super.acceptPyVisitor(pyVisitor)
+        }
+    }
+}